AI_service/requirements.txt
tensorflow
opencv-python
numpy
pandas
matplotlib
scikit-learn
pillow
protobuf
h5py
fastapi 
uvicorn 
python-multipart